// frontend/src/app/openapi/model/toolmodel-status.ts
/*
 * SPDX-FileCopyrightText: Copyright DB InfraGO AG and contributors
 * SPDX-License-Identifier: Apache-2.0
 *
 * Capella Collaboration
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * Do not edit the class manually.
 + To generate a new version, run `make openapi` in the root directory of this repository.
 */

import { PipelineRunStatus } from './pipeline-run-status';
import { ModelArtifactStatus } from './model-artifact-status';
import { GitModelStatus } from './git-model-status';


export interface ToolmodelStatus { 
    project_slug: string;
    model_slug: string;
    warnings: Array<string>;
    primary_git_repository_status: GitModelStatus;
    pipeline_status: PipelineRunStatus | null;
    model_badge_status: ModelArtifactStatus;
    diagram_cache_status: ModelArtifactStatus;
}
export namespace ToolmodelStatus {
}